[0.01] Move container factory resolution into application
